requirements.txt
fastapi
uvicorn
pydantic
motor
asyncpg
sqlalchemy
celery
redis
python-dotenv
asgiref
psycopg2-binary